refactor(gateway): extract audit_emission.go from server.go
Phase 3b file split: moves extractRRReconstructionFields, EmitConfigReloadAudit,
emitSignalReceivedAudit, emitSignalDeduplicatedAudit, emitCRDCreatedAudit,
retryAuditObserver/OnRetryAttempt, emitCRDCreationFailedAudit, and
constructReadableCorrelationID (DD-AUDIT-003 event emission) into their own
file. Pure structural move: 154 Ginkgo specs still pass.
